requirements.txt
python-dotenv
PyPDF2
langchain
langchain-openai
faiss-cpu
altair
streamlit